fixing diouxx len and prec

+++ includes/Libft/libft.h
@@ -20,7 +20,7 @@
 **	Conversion
 */
 int						ft_atoi(const char *str);
-char					*ft_itoa(int n);
+char					*ft_itoa(intmax_t n);
 char    				*ft_itoab_unsigned(uintmax_t num, int base);
 char    				*ft_itoab_signed(intmax_t num, int base);
 char    				*ft_ftoa(long double num, int precision);
